+++ src/styles/skin.ts
@@ -18,6 +18,7 @@ import RyniaTheme from './skins/rynia';
 import NiyaTheme from './skins/nyiaibu';
 import LumiTheme from './skins/lumi';
 import KeroroTheme from './skins/keroro';
+import skinTemplate from './skins/template';
 
 // needs to enable top-level await; necessary for other modules to import current skin config
 const setting = await getSettings();
@@ -43,7 +44,7 @@ export const SkinMap: { [key: string]: () => any } = {
   阿布: () => NiyaTheme(),
   露米Lumi: () => LumiTheme(),
   蛙吹Keroro: () => KeroroTheme(),
-  Custom: () => customSkin ?? NoxTheme(),
+  Custom: () => skinTemplate(customSkin) ?? NoxTheme(),
 };
 
 export const skins = (key = setting.skin) => {
@@ -60,6 +61,7 @@ export const skins = (key = setting.skin) => {
   };
   const playerStyle = getSkin();
   return {
+    ...NoxTheme(),
     ...playerStyle,
     outerLayerBtn: { padding: 'unset' },
     CRUDBtn: {
@@ -77,7 +79,7 @@ export const skins = (key = setting.skin) => {
       width: '1.1em',
       height: '1.1em',
       paddingBottom: '2px',
-      color: playerStyle.colorTheme.playListIconColor,
+      color: playerStyle.colorTheme?.playListIconColor,
     },
 
     AddFavIcon: {
@@ -85,17 +87,17 @@ export const skins = (key = setting.skin) => {
         cursor: 'pointer',
       },
       width: '1em',
-      color: playerStyle.colorTheme.playListIconColor,
+      color: playerStyle.colorTheme?.playListIconColor,
     },
 
     DiskIcon: {
       minWidth: '36px',
     },
     buttonStyle: buttonStyle(
-      playerStyle.reactJKPlayerTheme.sliderColor,
-      playerStyle.desktopTheme,
+      playerStyle.reactJKPlayerTheme?.sliderColor ?? '',
+      playerStyle.desktopTheme ?? '',
     ),
-    ScrollBar: ScrollBar(playerStyle.colorTheme.scrollbarColor),
+    ScrollBar: ScrollBar(playerStyle.colorTheme?.scrollbarColor ?? ''),
   };
 };
